WIP currently, lots of important infrastructure is partially complete.
The skeleton of the SAS functionality is there, with the evselect_image function basically done, and the ability to run the commands on a local machine via subprocess in place.

Creating command stacks is baked in, as well as choosing whether or not calling a function actually runs the commands or just constructs them.

A new set of 'product' classes are being written, they probably won't need to be super complicated, but they will cover the allowable data products this implementation of SAS can create. For instance the image class will act as a wrapper for an astropy HDUlist/numpy dictionary, with added methods and abilities to parse the stderr of the SAS call. Spectra will be wrappers of PyXspec spec objects and etc.