Started to implement the new functions required for generating XMM spectra.
First step is to reliably detect the X-ray peak of a cluster, to use as a central point.

New utility function to generate an annular (or circular) mask as a numpy array, for arbitrary inner/outer radii as well as within a angular range.

New method to try and make a positive match to an appropriate region implemented, as well as the ability to make a mask out of that region, and exclude any other sources that may fall within the object region.

Horrible problem with astropy.fits.open (I think), where I sometimes get a range of what seem to be C errors - segfault, double linked list, etc. Tried everything I know to fix it, so will have to go to the experts.

This commit is of an unfinished peak finding code.