WIP - evselect_spectrum command implemented and generating files.
Still haven't defeated the segfaults I get from astropy.fits, but lots of other behind-the-scenes work has gone on.

Many methods added to the BaseSource class - mostly dealing with regions, most useful for spectrum generation is the one which returns a SAS expression describing source and background regions with interlopers removed.

A function to generate annular/circular masks in the form of numpy arrays has been added.

Methods to generate and retrieve image masks for a source have been added.