Some minor bug fixes, and the two base XSPEC scripts required for XGA to work.
xga_extract.tcl was inspired by the writefits script in XSPEC, and writes a fairly complete summary of an XSPEC run to a fits file. This includes fit parameters, plotting information, and information about the individual spectra that go into the fit.

general_xspec_fit.xcm should be a completely general XSPEC script. It needs to be formatted in Python before it can run, but once it is it sets up the model, reads in data, sets up model paramaters, and calls xga_extract.tcl.