Change in behaviour of get_products.
In sources.py, the get_products method of a source instance will by default just return the product objects, rather than a list containing extra information like ObsID and instrument, as well as the product. The functions that relied on the old default behaviour have been updated to use the new behaviour.