Added a new method to setup a custom region in the ExtendedSource class.
Will either construct the region centered on the initial coordinates provided by the user, or iteratively search for the X-ray centroid.

Adapted _generate_mask to work with the new custom regions.

Added a new error for when iterative peak finding doesn't converge.

Added a counterpart to the the rad_to_ang function, ang_to_rad.

Added an option to the view method of the Image product class to allow the user to pass a mask.