Done some work toward PSF deconvolution, introduced a very bare bones PSF product class, that is a subclass of Image.
PSF class has a special get method that helps to resample the PSFs that SAS makes to something more useful for our images.

The codes that checks for the presence of XSPEC and SAS on the system have been moved from utils.py to xspec.py and sas.py respectively. Allows for convenience functions to be used without those pieces of software.

Added a psf coordinates attribute to sources.py, to store where a particular psf is generated at.

Started implementing a psfgen wrapper, though currently unfinished.