Fixed a stupid bug where ann_radii was calculating radii that were nonsense - due to a WCS thing.
The main fix was adding a new tool to imagetools 'pix_deg_scale', that calculates a multiplicative factor to change pixel radii to degrees. Had to alter ann_radii around that slighty.

Brightness profiles are now produced with a log scale on the x axis.

Radial and pizza profiles now have an extra mask applied if the product is a RateMap rather than an image - a detector map that weights pixels down to 0 if they aren't on the detector (so incorporates chip gaps too). Should make the brightness profiles more accurate.

Added a couple of properties to RateMap class to more easily access the masks it makes.

PSF functionality still under development