Contributes to issue #91, greatly speeds up the definition of GalaxyCluster source objects.
Went into the Image and RateMap classes and re-jigged how and when data is loaded in. WCS loading has been given its own on demand method, and only reads in the header info.

The way RateMaps work has also been altered, they now have a method that constructs the ratemap array on demand, so the image and exposure map data is only read in if the ratemap data is going to be used. Means I no longer waste time reading in a bunch of individual images/exposure maps I'll never use.

This all also dramatically improves the amount of memory I use.