Made huge changes in structure - parts of the radial_data_stack function have been split out into their own XGA internal functions (because they are used in the new radial_model_stack function). The same thing has happened in view_radial_data_stack.
radial_model_stack and view_radial_model_stack have been added. radial_model_stack allows the user to choose which model to fit to the profiles, then uses that to measure brightness values at self-similar scaled radii. view_radial_model_stack calls that function, but also adds a view of the fitted model and data if the show_images option is set to true.

All of the standard profile fitting options are available to pass into radial_model_stack - setting the number of steps, walkers etc for mcmc, or choosing to use a different fit method.

Closes issue #276