single_temp_apec has now been adapted to work with the new way we address spectra, in terms of their inner and outer radii and the SAS options given on creation
A huge number of other functions have been adapted to this as well, including storing and accessing fit results, and the functions that use those functions.

I also changed the rule on what a good temperature is in samples.extended.ClusterSample - now anything above 30keV is considered a failed fit. This closes issue #300