ProjectedGasTemperature1D now supports taking an upper limit as a keyword parameter - so illegally high values of temperature can be ignored.
Added properties to ProjectedGasTemperature1D to provide the user that upper limit, as well as an unedited version of the input y values if they want.

ProjectedGasTemperature1D and ProjectedGasMetallicity1D now take set_id as an input, so they can be aware of the spectrum they were generated from.