The inv_abel_fitted_model density function has been completely revamped, taking into account the new way of doing models. The inv_abel_data function has been removed, as it is no longer relevant or useful.
The norm_conv_factor has method has been changed slightly so that it returns the correct units of cm^-5/s.

The profile fit method now checks that (if a model instance is passed in rather than a name), that the model is set up with the correct units. If only a name is passed it makes sure to up a model with the correct units.

Some minor bug fixes have been applied to the GasDensity1D method